Replaced the C-style casts to C++-style casts
Since the using the C++ compiler, then the logical follow to the C++
style code, also removed some unnecessary casts.

+++ src/serialportengine_win.cpp
@@ -709,7 +709,7 @@ bool WinSerialPortEngine::setRate(qint32 rate, SerialPort::Directions dir)
         dptr->setError(SerialPort::UnsupportedPortOperationError);
         return false;
     }
-    m_currentDcb.BaudRate = DWORD(rate);
+    m_currentDcb.BaudRate = rate;
     return updateDcb();
 }
 
@@ -722,7 +722,7 @@ bool WinSerialPortEngine::setRate(qint32 rate, SerialPort::Directions dir)
 */
 bool WinSerialPortEngine::setDataBits(SerialPort::DataBits dataBits)
 {
-    m_currentDcb.ByteSize = BYTE(dataBits);
+    m_currentDcb.ByteSize = dataBits;
     return updateDcb();
 }
